test: merge overlay autofocus and focus-trap tests into focus.test.js
